requirements.txt
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
google-genai
python-dotenv
python-docx

# new (search + cleaner)
tavily-python
beautifulsoup4
readability-lxml
lxml_html_clean

# Google Agent Development Kit
google-adk
fastapi
uvicorn
